+++ ft_printf/printf_hex.c
@@ -6,13 +6,13 @@
 /*   By: hyudai <hyudai@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/05 19:42:42 by hyudai            #+#    #+#             */
-/*   Updated: 2021/02/12 17:47:39 by hyudai           ###   ########.fr       */
+/*   Updated: 2021/02/12 18:47:20 by hyudai           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-char	*hex_large(unsigned long num, int l, char *answer)
+char	*hex_large(unsigned int num, int l, char *answer)
 {
 	int		j;
 	const char			henkan[17] = "0123456789ABCDEF";
@@ -27,7 +27,7 @@ char	*hex_large(unsigned long num, int l, char *answer)
 	return (answer);
 }
 
-char	*hex_small(unsigned long num, int l, char *answer)
+char	*hex_small(unsigned int num, int l, char *answer)
 {
 	unsigned int	j;
 	const char			henkan[17] = "0123456789abcdef";
@@ -42,9 +42,9 @@ char	*hex_small(unsigned long num, int l, char *answer)
 	return (answer);
 }
 
-char	*hex_pointer(unsigned long num, int l, char *answer)
+char	*hex_pointer(unsigned long long num, int l, char *answer)
 {
-	unsigned long		j;
+	unsigned long long	j;
 	const char			henkan[17] = "0123456789abcdef";
 
 	j = 0;
